Change ResponsiveMatrixDemo to ResponsivePlayground and move to views
