+++ packages/client/webapp/src/router/index.ts
@@ -1,13 +1,13 @@
 import { createRouter, createWebHashHistory } from 'vue-router';
 import Home from '@/views/Home.vue';
-import ResponsiveMatrixDemo from '@/components/ResponsiveMatrixDemo.vue';
+import ResponsivePlayground from '@/views/ResponsivePlayground.vue';
 import DataExplorer from '@/views/DataExplorer.vue';
 import TA2Playground from '@/views/TA2Playground.vue';
 import Theia from '@/views/theia.vue';
 
 const routes = [
 	{ path: '/', component: Home },
-	{ path: '/results', component: ResponsiveMatrixDemo },
+	{ path: '/results', component: ResponsivePlayground },
 	{ path: '/explorer', component: DataExplorer },
 	{ path: '/ta2-playground', component: TA2Playground },
 	{ path: '/theia', component: Theia }
